Re-organize codebase
